+++ core/src/module.rs
@@ -117,9 +117,9 @@ impl Module {
     pub fn virtual_address(&self, file_offset: u64) -> Result<VA> {
         self.sections
             .iter()
-            .find(|&sec| sec.physical_range.contains(&(file_offset as u64)))
+            .find(|&sec| sec.physical_range.contains(&(file_offset)))
             .map(|sec| {
-                let section_offset = file_offset as u64 - sec.physical_range.start;
+                let section_offset = file_offset - sec.physical_range.start;
                 sec.virtual_range.start + section_offset
             })
             .ok_or_else(|| NotMapped.into())
